# feat: implement Phase 2 SWR optimization with 68-95% performance improvements