--- a/hooks/useFetchPacientes.ts
+++ b/hooks/useFetchPacientes.ts
@@ -12,9 +12,25 @@ export const useFetchPacientes = () => {
     "/pacientes/",
     fetcher,
     {
-      revalidateOnFocus: false, // Não revalidar quando a aba/janela ganhar foco
-      revalidateIfStale: false, // Não revalidar dados antigos automaticamente
-      dedupingInterval: 10000, // Deduplicar requisições similares em um intervalo de 10 segundos
+      revalidateOnFocus: false,
+      revalidateIfStale: true,
+      revalidateOnMount: true,
+      dedupingInterval: 25000, // 25 segundos - pacientes têm atualizações moderadas
+      refreshInterval: 180000, // 3 minutos - dados de pacientes são mais dinâmicos
+      keepPreviousData: true,
+      errorRetryCount: 3,
+      errorRetryInterval: 2000,
+      // Comparação otimizada para pacientes
+      compare: (a, b) => {
+        if (!a && !b) return true;
+        if (!a || !b) return false;
+        if (a.length !== b.length) return false;
+        return a.every(
+          (item, index) =>
+            item.id === b[index]?.id &&
+            item.updated_at === b[index]?.updated_at,
+        );
+      },
     },
   );
   return {